docker-etl-pipeline/etl/requirements.txt
psycopg2-binary==2.9.6
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0
